wined3d: Keep track of the onscreen depth stencil in the command stream instead of the device.
Based on a patch by Henri Verbeet.

TODO: Not sure if its the right way to go. There are certainly some
missing bits, e.g. accessing the cs via the device in all but one places
defeats the point of this at the moment.

Since the GL onscreen depth stencil is part of the WGL framebuffer,
which belongs to the swapchain, storing this info in the swapchain might
be the right thing to do. This should help multi-window or multi-head
configurations.