wined3d: Don't lock the src volume in device_update_volume
This will attempt to wait for the CS, but this code is executed by the CS, so
it will, among other problems, break the single producer, single consumer
assumption of the work queue.

FIXME: Maybe merge this with the previous patch or change their order.

FIXME 2: Loading the src location might be nice, or even better,
responding to unexpected src locs. In theory it should be in SYSMEM
since its a sysmem texture.