wined3d: Ignore WINED3D_MAP_NO_DIRTY_UPDATE in resource_map.
TODO: This flag and add_dirty_rect need tests.

The primary purpose of this patch is to fix a memory leak in World of
Warcraft. WoW uses WINED3D_MAP_NO_DIRTY_UPDATE | WINED3D_MAP_DISCARD on
sysmem surfaces. A new block of memory is allocated, but never assigned
to heap_mem because wined3d_cs_emit_resource_changed is not called.

The bigger picture is that we don't know how this flag and AddDirtyRect
/ AddDirtyBox are supposed to work. Msdn mentions some interaction with
update_surface and update_texture, but personally I think it is more
likely that this functionality is used to update separate rectangles in
a texture, similar to what can be achieved by mapping a buffer twice.