wined3d: Set map_heap_memory = NULL when allocating a PBO.
Otherwise the next unmap will re-assign it with the resource_changed
message. Freeing the actual memory and setting the read pointer to 0 is
the task of the worker thread.

Wined3d_resource_prepare_bo is only called if the map binding is
LOCATION_BUFFER, and map_heap_memory is only used if the map binding is
LOCATION_SYSMEM.