wined3d: Don't try to sync VBOs manually on OSX with CSMT.
This fixes a performance regression in some games caused by
aaf4849718c3c3a763c460a24a51dd6c6b55524a.

GL_ARB_sync and GL_APPLE_fence seem to be fairly slow on OSX. Putting the
buffer back into synchonized mode and synchronizing all future uploads
is about twice as fast. This was benchmarked with 3DMark2001. Gunpoint,
another game that depends on synchronized dynamic buffer maps, feels
smoother, although the lack of a benchmark mode makes testing this difficult.

This assumes that if a game maps a dynamic buffer without NOOVERWRITE or
DISCARD once it will always do so. I have not seen a game yet that mixes
async and sync maps. If the game mixes them, the wined3d-side CS will still
update the buffer asynchronously. Only the worker thread -> GL copy is
synchronized, as it used to be for all buffers in cx13.

This patch only changes the double-buffered (used for CSMT) codepath. Making
the same change to the glMapBuffer codepath improves 3DMark2001 performance
as well, but the difference is much smaller. I will leave the manual sync
code in place until we have a game that clearly profits from this change.