winemenubuilder: Create desktop shortcuts with absolute wine path.
When having multiple wine versions installed (for example regular wine and wine-compholio), the desktop
shortcuts will always run regular wine, even if the app was installed with wine-compholio. This patch
changes the behaviour to use the absolute wine path in desktop shortcuts. The patch only modifies the
behaviour on Linux, because some other distros are a bit special (FreeBSD requires a wrapper to start
wine, and so on ...).