ws2_32: Try harder to get the host name address in getaddrinfo()
When the host name is not resolvable getaddrinfo/GetAddrInfoW will
fail, this is not expected for some applications like League of
Legends [1][2][3]. We can deal with this in two ways:
- Try harder and use a NULL name to resolve the localhost address
(user transparent).
- Just warn the user and give up (requires the user to understand the
issue and fix /etc/hosts).

This patch tries harder and at the same time warns the user.

Tested on PC-BSD 9 and Debian 7 by removing the host name in /etc/hosts.

Fixes bug https://bugs.winehq.org/show_bug.cgi?id=29609 for some people.

[1] https://bugs.winehq.org/show_bug.cgi?id=29609#c10
[2] https://bugs.winehq.org/show_bug.cgi?id=29609#c13
[3] http://www.playonlinux.com/fr/topic-10056-League_of_Legends_crash_after_champ_select.html