dsound: Fix angle to sound source calculation.
This patch fixes incorrect sound positioning in dsound.
As of now, angle to sound source is calculated as angle between
vDistance and vOrientFront.
This leads to incorrect results for all sources that are not in the
same plane as speakers.
In extreme case: for sources directly above or below origin, angle is
calculated to -90 degrees, and they are played in the left speaker.

This patch changes angle calculation: angle to sound source is
calculated as angle between vDistance and plane given by vectors
vOrientFront and vOrientTop.