ucrtbase: Add the new universal CRT DLL.
Since MSVC 2015/Windows 10, the C runtime has now been split into two
parts, ucrtbase, which is the generic C runtime which is now considered
a system component (in Windows 10, and as an extra installed runtime
component for older Windows version), and vcruntimeX which is specific
to the compiler version.

This uses the msvcrt implementation, just like the earlier msvcr* DLLs.

Functions with names that existed in msvcr120 are hooked up so far.