widl: When adding an interface typedef do check whether it has been already added while resolving the parent interface.
This patch fixes a long standing bug that when an inherited interface gets
referenced by a parent interface definition then multiple definitions of an
inherited interface get added to a typelib.

This bug has been discovered by the patch that emits a warning for a duplicate
uuid. A test case for this problem is added in the last patch of the series.

Signed-off-by: Sebastian Lackner <sebastian@fds-team.de>
Signed-off-by: Dmitry Timoshkov <dmitry@baikal.ru>