ntdll: Block signals while executing system APCs. (v2)
Changes in v2:
  * Do not delete the still valid comment "Return TRUE if a user APC has been run.".
    I still had this deleted from an earlier version of the patchset, where the
    return type was changed.

If a USR1 suspend signal arrives between dequeing a system APC and sending
back the result to the wineserver, a deadlock occurs. To avoid this issue
this patch blocks all signals in server_select(), except while waiting or
processing user APCs.

In server_queue_process_apc() we do not need to block signals because no
other process will wait on the result.

Signed-off-by: Sebastian Lackner <sebastian@fds-team.de>