winex11.drv: Allow to select default display frequency in registry key.
When an application doesn't request a specific display frequency, Wine
currently just picks the first one. Most of the time this is fine, but
there is no way to switch to a different frequency. This patch adds a
registry key (of type STRING) to select the default display frequency:

HKCU\Software\Wine\X11 Driver\DefaultDisplayFrequency