shell32: Check IsWoW64Process before calling Wow64 functions (try 2)
Replace patch entry 116937.

Fix one of the error leak source that prevents the installer to succeed
(bug 36838).

On 32 bit Wine, the original code did set last error to
ERROR_CALL_NOT_IMPLEMENTED when calling wow64 functions. The game
installer chokes at some point when the last error is anything but zero
and fails to complete the installation. Checking if wow64 is available
before calling wow64 functions avoids that. The affected wow64 function
calls themselves are not required for proper operation on 32 bit
systems.

Signed-off-by: Olivier F. R. Dierick <o.dierick@piezo-forte.be>