user32: Do not initialize dialog info for every window passed to DefDlgProc.
The tests in the last patch in the series use
DefDlgProcA(parent, DM_GETDEFID, 0, 0);
as a simple dialog detector, and that call leads to dialog info being
created for a not dialog window, which later makes this window being
treated as a dialog.

DIALOG_CreateControls32() gets called after the dialog info creation
has been forced, so passing TRUE to DIALOG_get_info() is redundant there.